requirements.txt
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
openai==0.8.0
python-dotenv==0.21.1
python-telegram-bot==13.7
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
